--amend

+++ src/neural_astar/planner/differentiable_astar.py
@@ -115,7 +115,7 @@ def backtrack(
     """
 
     num_samples = start_maps.shape[0]
-    parents = parents.type(int)
+    parents = parents
     goal_maps = goal_maps.type(torch.long)
     start_maps = start_maps.type(torch.long)
     path_maps = goal_maps.type(torch.long)
